- [Graphics] Added MeshLoader Class for assimp loading of fbx - [Graphics] Added Load Mesh overload function to load a pre existing mesh - [Angazi] Updated Material and mesh components to read in from assimp - [Assets] Added 10k images for earth
git-tfs-id: [https://dev.azure.com/timjustin/]$/Angazi;C396

+++ Framework/Graphics/Src/AnimationIO.cpp
@@ -81,7 +81,7 @@ void AnimationIO::Write(FILE * file, const AnimationClip & animationClip)
 void AnimationIO::Read(FILE * file, AnimationClip & animationClip)
 {
 	char buffer[128];
-	fscanf_s(file, "Name: %s\n", buffer, std::size(buffer));
+	fscanf_s(file, "Name: %s\n", buffer, (uint32_t) std::size(buffer));
 	animationClip.name = buffer;
 
 	fscanf_s(file, "Duration: %f\n", &animationClip.duration);
@@ -92,7 +92,7 @@ void AnimationIO::Read(FILE * file, AnimationClip & animationClip)
 	animationClip.boneAnimations.resize(totalAnimations);
 	for (uint32_t i = 0; i < totalAnimations; i++)
 	{
-		fscanf_s(file, "%s\n", buffer, std::size(buffer));
+		fscanf_s(file, "%s\n", buffer, (uint32_t) std::size(buffer));
 		if (std::string(buffer) == "[Animation]")
 		{
 			animationClip.boneAnimations[i] = std::make_unique<Animation>();
